SM-4 Move FileSplitter into the [Application] Module.
