+++ Applications/src/main/java/com/oliver/FileSplitter.java
@@ -1,10 +1,12 @@
-package com.util;
+package com.oliver;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 
@@ -20,7 +22,8 @@ public class FileSplitter {
     log.info("end");
   }
 
-  static void splitFileInSameFolder(Path sourceFilePath, int partitions) {
+  static List<Path> splitFileInSameFolder(Path sourceFilePath, int partitions) {
+    assert partitions > 0;
     long lineCount = countLineJava8(sourceFilePath);
     log.debug("lineCount={}", lineCount);
     long step = (lineCount + partitions) / partitions;
@@ -31,6 +34,7 @@ public class FileSplitter {
       long end = Math.min(begin + step, lineCount - 1);
       log.debug("r={}, begin={} -> end={}", r, begin, end);
     }
+    List<Path> chunkFiles = new ArrayList<>();
     try (BufferedReader reader = Files.newBufferedReader(sourceFilePath)) {
       for (r = 0; r < partitions; r++) {
         long begin = r * step;
@@ -39,10 +43,12 @@ public class FileSplitter {
         try (BufferedWriter writer = Files.newBufferedWriter(targetPath)) {
           readWriteChunkOfLines(reader, writer, end - begin + 1);
         }
+        chunkFiles.add(targetPath);
       }
     } catch (IOException e) {
       log.error(e.getMessage());
     }
+    return chunkFiles;
   }
 
   static void readWriteChunkOfLines(BufferedReader reader, BufferedWriter writer, long numOfLines) throws IOException {
