+++ Applications/src/test/resources/TestFile.txt
@@ -1,0 +1,5 @@
+1
+2
+3
+4
+5
